% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warping.R
\name{bernstein_warping_design_matrix}
\alias{bernstein_warping_design_matrix}
\title{Construct a Bernstein Warping Design Matrix}
\usage{
bernstein_warping_design_matrix(x, order, domain)
}
\arguments{
\item{x}{Numeric vector representing the coordinate values; must be within
the given domain.}

\item{order}{An integer representing the order of the Bernstein polynomial.}

\item{domain}{A numeric vector of length 2 specifying the domain over which
the Bernstein polynomial is defined.}
}
\value{
A matrix where each column represents a Bernstein polynomial
evaluated at the data points \code{x}.
}
\description{
Creates a design matrix based on the Bernstein basis for use in GeoWarp's
axial warping units. The difference to the usual Bernstein basis is that the
first basis function is omitted so that input coordinate zero maps to warped
coordinate zero. The order determines the flexibility of the basis expansion.
}
\examples{
# Create a Bernstein warping design matrix with 10 data points, of order 4,
# over domain [0, 5]
design_matrix <- bernstein_warping_design_matrix(
  x = seq(0, 5, length.out = 10),
  order = 4,
  domain = c(0, 5)
)
print(design_matrix)

}
