% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{geowarp_cross_validation_fit}
\alias{geowarp_cross_validation_fit}
\title{Perform Cross-Validation for GeoWarp}
\usage{
geowarp_cross_validation_fit(
  df,
  grouping_variable,
  model,
  method = "geowarp_optimise",
  show_progress = FALSE,
  grouping_cores = 1,
  group_output_pattern,
  prepare_input = function(input_df, group) input_df,
  ...
)
}
\arguments{
\item{df}{Data frame containing the observations.}

\item{grouping_variable}{Variable that defines the grouping in \code{df}.}

\item{model}{A GeoWarp model made using \code{\link{geowarp_model}} that
specifies the model structure and priors.}

\item{method}{The method used for fitting the model, defaults to
\code{\link{geowarp_optimise}}.}

\item{show_progress}{Whether to show progress, defaults to FALSE.}

\item{grouping_cores}{Number of fits to run in parallel.}

\item{group_output_pattern}{A \code{\link[glue]{glue}} pattern for naming
output files used for logging output (optional).}

\item{prepare_input}{A function to prepare the input data for each group,
defaults to identity function.}

\item{...}{Additional arguments to be passed to the method for fitting.}
}
\value{
A list containing the cross-validated fits for each group.
}
\description{
This function performs k-fold cross-validation on a specified dataset and
GeoWarp model. Each group defined by \code{grouping_variable} is left out once
while fitting the model with the remaining data. The function returns a list
of fitted models for each left-out group.
}
\examples{
# Using geowarp_cross_validation_fit
cv_fits <- geowarp_cross_validation_fit(
  df = my_data,
  grouping_variable = 'group_id',
  model = my_model,
  show_progress = TRUE
)

}
