% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{geowarp_deviation_model}
\alias{geowarp_deviation_model}
\alias{geowarp_white_deviation_model}
\alias{geowarp_vertical_only_deviation_model}
\title{Create a GeoWarp Deviation Model Object}
\usage{
geowarp_deviation_model(
  covariance_function = c("matern15", "exponential"),
  axial_warping_units,
  axial_warping_unit_mapping = seq_len(length(axial_warping_units)),
  geometric_warping_unit = geowarp_geometric_warping_unit(),
  variance_model = geowarp_variance_model()
)

geowarp_white_deviation_model(variance_model = geowarp_variance_model())

geowarp_vertical_only_deviation_model(
  covariance_function = c("matern15", "exponential"),
  axial_warping_unit = geowarp_bernstein_awu(),
  variance_model = geowarp_variance_model()
)
}
\arguments{
\item{covariance_function}{A character vector specifying the type of
covariance function to use. Options are 'matern15' and 'exponential'.
Defaults to 'matern15'.}

\item{axial_warping_units}{A list of axial warping unit objects that specify
the warping functions for all the coordinates (for
\code{geowarp_deviation_model}). One of \code{\link{geowarp_linear_awu}}, or
\code{\link{geowarp_bernstein_awu}}. At the moment, horizontal coordinates
must only have linear warpings.}

\item{axial_warping_unit_mapping}{Specifies the mapping between the axial
warping units and the coordinates. Can be used to share warping units between
coordinates.}

\item{geometric_warping_unit}{A geometric warping unit object. Defaults to
\code{geowarp_geometric_warping_unit()}.}

\item{variance_model}{A variance model object. Defaults to
\code{geowarp_variance_model()}.}
}
\value{
A list containing the specifications for the GeoWarp deviation model,
ready to be passed as an argument to \code{\link{geowarp_model}}.
}
\description{
Constructs a GeoWarp deviation model object specifying the correlation
structure, the axial warping units, and the variance model. This object is
intended to be used as an argument for \code{\link{geowarp_model}}.
}
\section{Functions}{
\itemize{
\item \code{geowarp_white_deviation_model()}: A deviation model that assumes no
correlation (that is, i.i.d. Gaussian)

\item \code{geowarp_vertical_only_deviation_model()}: A deviation model that assumes no
correlation in the horizontal directions.

}}
\examples{
# Creating a GeoWarp deviation model object with custom settings
deviation_model <- geowarp_deviation_model(
  covariance_function = 'exponential',
  axial_warping_units = list(geowarp_linear_awu(), geowarp_bernstein_awu())
)

}
\seealso{
\code{\link{geowarp_model}}, \code{\link{geowarp_mean_model}},
\code{\link{geowarp_linear_awu}}, \code{\link{geowarp_bernstein_awu}},
\code{\link{geowarp_geometric_warping_unit}},
\code{\link{geowarp_variance_model}}
}
