% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{geowarp_linear_awu}
\alias{geowarp_linear_awu}
\alias{geowarp_bernstein_awu}
\alias{geowarp_geometric_warping_unit}
\title{Create a GeoWarp Axial Warping Unit (AWU) Object}
\usage{
geowarp_linear_awu(scaling = 1, prior = list(shape = 1.01, rate = 0.01))

geowarp_bernstein_awu(order = 20, prior = list(shape = 1.01, rate = 0.01))

geowarp_geometric_warping_unit(prior_shape = 2)
}
\arguments{
\item{scaling}{Numeric value specifying the scaling for the linear warping
unit. This is used to rescale the input coordinate, and the specified prior
applies after rescaling. This help with numerical issues by making the
inferred parameter closer to one.}

\item{prior}{A list giving the parameters for the gamma prior over the
axial warping unit parameters. A list containing two entries, \code{shape} and
\code{rate}.}

\item{order}{Numeric value specifying the order of the Bernstein polynomial.
Defaults to 20.}

\item{prior_shape}{Numeric value specifying the shape parameter for the
prior for the geometric warping unit. Defaults to 2. The value 1 implies a
uniform prior over the warping, while values greater than 1 imply shrinkage
towards no warping.}
}
\value{
A list representing the warping unit, ready to be passed as an
argument to \code{\link{geowarp_deviation_model}}.
}
\description{
Constructs a GeoWarp warping unit object for use in the deviation model
(see \code{\link{geowarp_deviation_model}}).
}
\details{
Axial warping units operate on individual coordinates, and can be either
linear (\code{\link{geowarp_linear_awu}}) or non-linear using a Bernstein
polynomial (\code{\link{geowarp_bernstein_awu}}). A geometric warping unit
induces geometric anisotropy in the space, and is specified through
\code{\link{geowarp_geometric_warping_unit}}.
}
\section{Functions}{
\itemize{
\item \code{geowarp_bernstein_awu}: A warping unit that uses Bernstein
polynomials.

\item \code{geowarp_geometric_warping_unit}: A geometric warping unit that specifies
geometric anisotropy.
}}

\examples{
# Creating a GeoWarp linear AWU
linear_awu <- geowarp_linear_awu(scaling = 0.5)

# Creating a GeoWarp Bernstein AWU
bernstein_awu <- geowarp_bernstein_awu(order = 10)

# Creating a GeoWarp geometric warping unit
geometric_wu <- geowarp_geometric_warping_unit(prior_shape = 3)

}
\seealso{
\code{\link{geowarp_model}}, \code{\link{geowarp_deviation_model}},
\code{\link{geowarp_variance_model}}
}
