% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{geowarp_mean_model}
\alias{geowarp_mean_model}
\title{Create a GeoWarp Mean Model Object}
\usage{
geowarp_mean_model(
  fixed_formula,
  fixed_effect_mean = 0,
  fixed_effect_precision = 1/100,
  vertical_basis_functions = TRUE,
  vertical_basis_function_delta = 0.1,
  vertical_basis_function_effect_precision = 1/100,
  vertical_basis_function_variance_prior = inverse_gamma_quantile_prior(0.001^2, 10^2)
)
}
\arguments{
\item{fixed_formula}{A formula or character vector that specifies the fixed
effects. If a character vector is provided, it will be assumed to refer to
a single variable and converted to a formula.}

\item{fixed_effect_mean}{A numeric value representing the mean of the fixed
effect. Defaults to 0.}

\item{fixed_effect_precision}{A numeric value representing the precision of
the fixed effect. Defaults to 1 / 100.}

\item{vertical_basis_functions}{Logical value indicating whether to include
vertical basis functions. Defaults to TRUE.}

\item{vertical_basis_function_delta}{A numeric value that specifies the
spacing of the vertical basis functions. Defaults to 0.1.}

\item{vertical_basis_function_effect_precision}{A numeric value representing
the precision of the effect of the vertical basis functions. Defaults to
1 / 100.}

\item{vertical_basis_function_variance_prior}{Prior distribution for the
variance of the vertical basis functions, given as a list with \code{shape} and
\code{rate} parameters. Defaults to an inverse gamma prior with 5\% and 95\%
percentiles of \code{1e-3^2} and \code{10^2} respectively}
}
\value{
A GeoWarp mean model, ready to be passed as an argument to
\code{\link{geowarp_model}}.
}
\description{
Constructs a GeoWarp mean model object that specifies the fixed effects
formula, the mean and precision of the fixed effects, and parameters for
vertical basis functions. This object is intended to be used as an argument
for \code{\link{geowarp_model}}.
}
\examples{
# Creating a GeoWarp mean model object with custom settings
mean_model <- geowarp_mean_model(
  fixed_formula = ~ depth,
  fixed_effect_mean = 1,
  fixed_effect_precision = 1,
  vertical_basis_function_delta = 0.2
)

}
\seealso{
\code{\link{geowarp_model}}, \code{\link{geowarp_deviation_model}}
}
