% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{geowarp_model}
\alias{geowarp_model}
\title{Construct a GeoWarp Model Object}
\usage{
geowarp_model(
  variable,
  horizontal_coordinates,
  vertical_coordinate,
  horizontal_domains,
  vertical_domain,
  mean_model = geowarp_mean_model(fixed_formula = vertical_coordinate),
  deviation_model = geowarp_deviation_model(axial_warping_units =
    c(rep(list(geowarp_linear_awu()), length(horizontal_coordinates)),
    list(geowarp_bernstein_awu()))),
  nugget_prior = inverse_gamma_quantile_prior(0.1, 1)
)
}
\arguments{
\item{variable}{Character vector naming the target variable.}

\item{horizontal_coordinates}{Vector containing names of horizontal
coordinates.}

\item{vertical_coordinate}{Vector containing name of vertical coordinate.}

\item{horizontal_domains}{Domain of horizontal coordinates, or a list if
there is more than one. A domain is a numeric vector of length two.}

\item{vertical_domain}{Domain of the vertical coordinate, a vector of length
two.}

\item{mean_model}{GeoWarp mean model object. Defaults to the model given in
the GeoWarp paper, with an intercept and a trend based on the vertical
coordinate. See \code{\link{geowarp_mean_model}} for details.}

\item{deviation_model}{GeoWarp deviation model object. Defaults match the
configuration given in the GeoWarp paper.}

\item{nugget_prior}{Prior distribution for the nugget effect given as a list
with \code{shape} and \code{rate} parameters.}
}
\value{
A \code{geowarp_model} object containing the specifications for the
GeoWarp model, ready to be passed to \code{\link{geowarp_optimise}}.
}
\description{
Creates a GeoWarp model object that specifies the structure and priors of
the model. The created object is intended to be used as input for
\code{\link{geowarp_optimise}}.
}
\examples{
# Creating a simple GeoWarp model object for a 2-D problem
model <- geowarp_model(
  variable = 'log_q_c',
  horizontal_coordinates = 'horizontal',
  vertical_coordinate = 'depth',
  horizontal_domains = c(0, 150),
  vertical_domain = c(0, 43)
)
# Creating a simple GeoWarp model object for a 3-D problem
model <- geowarp_model(
  variable = 'log_q_c',
  horizontal_coordinates = c('easting', 'northing'),
  vertical_coordinate = 'depth',
  horizontal_domains = list(c(0, 150), c(0, 200)),
  vertical_domain = c(0, 43)
)
# Overriding some of the parameters, namely the spacing of the splines used
# in both the mean model and the depth-varying variances
model <- geowarp_model(
  variable = 'log_q_c',
  horizontal_coordinates = c('easting', 'northing'),
  vertical_coordinate = 'depth',
  horizontal_domains = list(c(0, 150), c(0, 200)),
  vertical_domain = c(0, 43),
  mean_model = geowarp_mean_model(
    fixed_formula = ~ depth,
    # Set the spacing for knots in the mean function to 1
    vertical_basis_function_delta = 1
  ),
  deviation_model = geowarp_deviation_model(
    axial_warping_units = list(
      geowarp_linear_awu(),
      geowarp_linear_awu(),
      geowarp_bernstein_awu()
    ),
    variance_model = geowarp_variance_model(
      # Set knot spacing to 5
      vertical_basis_function_delta = 5
    )
  )
)

}
\seealso{
\code{\link{geowarp_mean_model}}
\code{\link{geowarp_deviation_model}}
\code{\link{geowarp_optimise}}
}
