% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{geowarp_optimise}
\alias{geowarp_optimise}
\title{Estimate the GeoWarp Model Parameters Through Optimisation}
\usage{
geowarp_optimise(
  df,
  model = geowarp_model(),
  max_attempts = 20,
  best_of = 5,
  method = c("stan_nlminb", "stan_nlm", "stan_trust_optim", "optimizing"),
  vecchia = "auto",
  n_parents = 20,
  parent_structure = vecchia_parent_structure(df, model, n_parents, ),
  grouping = vecchia_grouping(parent_structure),
  threads = getOption("geowarp.threads"),
  grain_size = 100L,
  ...
)
}
\arguments{
\item{df}{Data frame containing the observations.}

\item{model}{A GeoWarp model object specifying the model structure and
priors,created using \code{\link{geowarp_model}}.}

\item{max_attempts}{Maximum number of optimisation attempts. Defaults to 20.}

\item{best_of}{Number of times to run the optimisation algorithm. The
algorithm is run repeatedly with different starting values in order to help
find the global posterior mode.}

\item{method}{Optimisation algorithm to use, one of either 'stan_trust_optim'
or 'optimizing'. The first, the default, uses the
\code{\link{stan_trust_optim}} function, while the second uses
\code{\link[rstan]{optimizing}}.}

\item{vecchia}{Whether to use the Vecchia approximation. Can be 'auto',
\code{TRUE}, or \code{FALSE}. If 'auto', the Vecchia approximation is used if there
are more than 1000 observations.}

\item{n_parents}{Number of parent locations for the Vecchia approximation.
Defaults to 20.}

\item{parent_structure}{The parent structure used for the Vecchia
approximation; see \code{\link{vecchia_parent_structure}}.}

\item{grouping}{Grouping structure for the Vecchia approximation; see
\code{\link{vecchia_grouping}}. The default should be fine for most cases.}

\item{threads}{Number of threads to use for parallelisation, taken from
\code{getOption('geowarp.threads')} by default, which is itself set to 1 by
default. The special value -1 picks a number of threads based on the number
of cores in the system).}

\item{grain_size}{Used to control the parallelisation. The default, 100,
should be okay. On systems with few cores, a larger value might be better.}

\item{...}{Additional parameters to be passed to the chosen optimisation
routine in \code{method}.}
}
\value{
A 'geowarp_fit' object that estimated parameters, suitable for use
with \code{\link{predict.geowarp_fit}} and other methods.
}
\description{
This function estimates the parameters of a specified GeoWarp model by using
an optimisation algorithm to find the posterior mode.
}
\examples{
# Using the geowarp_optimise function
fit <- geowarp_optimise(
  df = my_data,
  model = my_model
)

}
