% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{geowarp_simulate}
\alias{geowarp_simulate}
\title{Simulate from a GeoWarp Fit}
\usage{
geowarp_simulate(
  fit,
  df = fit$observed_df,
  model = fit$model,
  parameters = fit$parameters,
  n_samples = 1,
  nugget = TRUE,
  vecchia = "auto",
  vecchia_n_parents = 50,
  parent_structure = vecchia_parent_structure(df, model, vecchia_n_parents),
  threads = getOption("geowarp.threads")
)
}
\arguments{
\item{fit}{GeoWarp fit object made using \code{\link{geowarp_optimise}}. This
is optional if \code{df}, \code{model}, and \code{parameters} are given directly.}

\item{df}{A data frame containing the locations to simulated (defaults to the
observation locations)}

\item{model}{GeoWarp model object used for simulation (default uses the model
from the \code{fit} object).}

\item{parameters}{Parameters used for the simulation (default uses the
parameters from the \code{fit} object).}

\item{n_samples}{The number of samples to simulate (default is 1).}

\item{nugget}{Logical indicating whether to include a nugget effect (default
is TRUE).}

\item{vecchia}{Either 'auto', TRUE, or FALSE. If 'auto', vecchia
approximation is used for n > 1000 (default is 'auto').}

\item{vecchia_n_parents}{Number of parents in the Vecchia approximation
(default is 50).}

\item{parent_structure}{The parent structure for Vecchia approximation
constructed using \code{\link{vecchia_parent_structure}}. The default parent
structure assumes that the data are vertically dense, which is not
appropriate if they are uniform or on a grid; see the documentation for that
function.}

\item{threads}{Number of threads to use for parallelisation, taken from
\code{getOption('geowarp.threads')} by default, which is itself set to 1 by
default. The special value -1 picks a number of threads based on the number
of cores in the system).}
}
\value{
A copy of the input data frame \code{df} with tje simulated values added
as a new column.
}
\description{
This function performs unconditional simulation from a fitted GeoWarp model.
The simulation conditions only on the estimated parameters and not on the
data. To simulate conditioned on data, use \code{\link{predict.geowarp_fit}}.
}
\examples{
# Assuming `fit` is a fitted GeoWarp model:
simulated_df <- geowarp_simulate(fit)

}
