% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{geowarp_stan_model}
\alias{geowarp_stan_model}
\alias{geowarp_stan_data}
\title{GeoWarp Stan Model}
\usage{
geowarp_stan_model(name = c("full", "vertical_only", "white"))

geowarp_stan_data(
  df,
  model,
  parent_structure,
  grouping,
  white_block_size = min(nrow(df) - 1L, 30)
)
}
\arguments{
\item{name}{Name of the model to use. Can be 'full', 'vertical_only', or
'white'. The full model is the default.}

\item{df}{The input data to prepare for Stan.}

\item{model}{The GeoWarp model object, created using
\code{\link{geowarp_model}}.}

\item{parent_structure}{The parent structure used for the Vecchia
approximation; see \code{\link{vecchia_parent_structure}}.}

\item{grouping}{Grouping structure for the Vecchia approximation; see
\code{\link{vecchia_grouping}}. The default should be fine for most cases.}

\item{white_block_size}{Block size to use for the white noise model. This
affects performance but not the results; the default should be fine.}
}
\value{
A \link[rstan]{stan_model} object.
}
\description{
These functions allow access to the underlying Stan implementation of
GeoWarp. They allow access to the \link[rstan]{stan_model} object used to fit
GeoWarp, and can prepare the input data required to use the object.
}
\section{Functions}{
\itemize{
\item \code{geowarp_stan_data}: Prepare data for Stan
}}

\seealso{
\code{\link{geowarp_optimise}}
}
