% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{geowarp_variance_model}
\alias{geowarp_variance_model}
\title{Create a GeoWarp Variance Model Object}
\usage{
geowarp_variance_model(
  fixed_formula = ~1,
  fixed_effect_mean = 0,
  fixed_effect_precision = 1/100,
  vertical_basis_functions = TRUE,
  vertical_basis_function_delta = 1,
  vertical_basis_function_length_scale_prior = list(scale = 1),
  vertical_basis_function_variance_prior = inverse_gamma_quantile_prior(0.001^2, 10^2)
)
}
\arguments{
\item{fixed_formula}{A formula specifying the fixed effect. Defaults to an
intercept-only model (\code{~ 1}).}

\item{fixed_effect_mean}{Numeric value for the prior mean of the fixed
effect.}

\item{fixed_effect_precision}{Numeric value for the prior precision of the
fixed effect.}

\item{vertical_basis_functions}{Logical indicating whether to include
vertical basis functions. Defaults to \code{TRUE}.}

\item{vertical_basis_function_delta}{Numeric value specifying the spacing
between knots for the vertical basis functions. Defaults to 1.}

\item{vertical_basis_function_length_scale_prior}{List containing the prior
for the length scale parameter for the basis function coefficients. A list
containing one entry, \code{scale}.}

\item{vertical_basis_function_variance_prior}{A prior object for the variance
of the vertical basis functions, given as a list with \code{shape} and
\code{rate} parameters. Defaults to an inverse gamma prior with 5\% and 95\%
percentiles of \code{1e-3 ^ 2} and \code{10 ^ 2} respectively}
}
\value{
A list containing the specifications for the GeoWarp variance model,
ready to be passed as an argument to \code{\link{geowarp_deviation_model}}
and \code{\link{geowarp_model}}.
}
\description{
Constructs a GeoWarp variance model object that specifies the model for
the log variance of the deviation process. The log variance has a fixed
effect, and optional vertical basis functions. This object is intended to be
used as an argument for \code{\link{geowarp_deviation_model}} and
\code{\link{geowarp_model}}.
}
\examples{
# Creating a GeoWarp variance model object with custom settings
variance_model <- geowarp_variance_model(
  fixed_formula = ~ depth,
  vertical_basis_function_delta = 0.5
)

}
\seealso{
\code{\link{geowarp_model}}, \code{\link{geowarp_deviation_model}},
\code{\link{geowarp_mean_model}}
}
