% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{mean_profile}
\alias{mean_profile}
\alias{marginal_variance_profile}
\title{Calculate estimated depth-dependent mean and variance}
\usage{
mean_profile(
  fit,
  df = fit$observed_df,
  model = fit$model,
  parameters = fit$parameters
)

marginal_variance_profile(
  fit,
  df = fit$observed_df,
  model = fit$model,
  parameters = fit$parameters,
  nugget = TRUE
)
}
\arguments{
\item{fit}{A GeoWarp fit object form \code{\link{geowarp_optimise}}. Can be
omitted if both \code{model} and \code{parameters} are given.}

\item{df}{A data frame containing the observed data (default uses the
observation locations used in \code{fit}).}

\item{model}{GeoWarp model object (default uses the model from the \code{fit}
object).}

\item{parameters}{GeoWarp parameters (default uses the parameters from the
\code{fit} object).}

\item{nugget}{Logical indicating whether to include the nugget variance in
the output (default is TRUE).}
}
\value{
A vector containing the estimated mean or marginal variance at each
location in \code{df}.
}
\description{
In the GeoWarp model, the process mean and variance vary with depth. These
functions calculate the estimated mean and variance at the given locations.
}
\section{Functions}{
\itemize{
\item \code{marginal_variance_profile()}: Output the marginal variance at each depth

}}
\examples{
# Assuming `fit` is a fitted GeoWarp model:
mean_values <- mean_profile(fit)
}
