% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_coordinates}
\alias{model_coordinates}
\title{Produce Matrix of Coordinates for a GeoWarp Model}
\usage{
model_coordinates(df, model)
}
\arguments{
\item{df}{A data frame containing the coordinates.}

\item{model}{A GeoWarp model object made using \code{\link{geowarp_model}}.}
}
\value{
A matrix containing the coordinates extracted from the data frame
according to the specifications in the GeoWarp model object. The columns
are ordered according to \code{model$horizontal_coordinates} followed by
\code{model$vertical_coordinate}.
}
\description{
Extracts the coordinates specified in the given GeoWarp model from the input
data frame and returns them as a matrix.
}
\examples{
# Sample data frame
df <- data.frame(x = c(1, 2, 3), y = c(4, 5, 6), z = c(7, 8, 9))

# Sample GeoWarp model object
model <- geowarp_model(
  variable = 'value',
  horizontal_coordinates = c('x', 'y'),
  horizontal_domains = list(c(0, 10), c(0, 10)),
  vertical_coordinate = 'z',
  vertical_domain = c(0, 10)
)

# Generate matrix of coordinates
coordinates_matrix <- model_coordinates(df, model)

}
\seealso{
\code{\link{geowarp_model}}, \code{\link{geowarp_deviation_model}},
\code{\link{geowarp_variance_model}}
}
