% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcptmodels-package.R
\docType{package}
\name{pcptmodels-package}
\alias{pcptmodels-package}
\alias{pcptmodels}
\title{The 'pcptmodels' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.7. https://mc-stan.org
}
