% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia.R
\name{plot_parent_structure}
\alias{plot_parent_structure}
\title{Plot a Vecchia Parent Structure}
\usage{
plot_parent_structure(
  observed_df,
  model,
  structure,
  prediction_df,
  observed_indices,
  prediction_indices,
  left_coordinate = 1,
  right_coordinate = 2
)
}
\arguments{
\item{observed_df}{A data frame containing the observed data.}

\item{model}{GeoWarp model object.}

\item{structure}{A list containing parent structures and orderings produced
by \code{\link{vecchia_parent_structure}}.}

\item{prediction_df}{A data frame containing the prediction data (optional).}

\item{observed_indices}{Indices of observed data points to plot (optional;
defaults to three chosen at random).}

\item{prediction_indices}{Indices of prediction data points to plot
(optional; defaults to three chosen at random).}

\item{left_coordinate}{Coordinate to use for the plot x-axis (default is 1).}

\item{right_coordinate}{Coordinate to use for the plot y-axis (default is 2).}
}
\value{
A ggplot2 plot.
}
\description{
This function visualizes a Vecchia parent structure of data in a 2D plot.
Each panel in the plot is either the target, one of its parents, or other.
}
