% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_alpha_beta}
\alias{predict_alpha_beta}
\title{Give the posterior distribution of alpha beta}
\usage{
predict_alpha_beta(
  fit,
  df = fit$observed_df,
  parameters = fit$parameters,
  model = fit$model,
  include_mean = TRUE,
  include_precision = FALSE,
  include_samples = FALSE,
  n_samples = 1L,
  n_parents = 50,
  parent_structure = vecchia_parent_structure(df, model, n_parents),
  threads = getOption("geowarp.threads")
)
}
\arguments{
\item{fit}{A GeoWarp fit object from \code{\link{geowarp_optimise}}.}

\item{df}{A data frame containing the observed data (default uses the
observation locations used in \code{fit}).}

\item{parameters}{GeoWarp parameters (default uses the parameters from the
\code{fit} object).}

\item{model}{GeoWarp model object (default uses the model from the \code{fit}
object).}

\item{include_mean}{Logical, indicating whether to include the mean in the
output.}

\item{include_precision}{Logical, indicating whether to include the precision
matrix in the output (explained below).}

\item{include_samples}{Logical, indicating whether to include the predictive
samples in the output.}

\item{n_samples}{Number of predictive samples to generate.}

\item{n_parents}{Number of parents to consider for Vecchia's approximation.}

\item{parent_structure}{Parent structure for Vecchia's approximation,
produced using \code{\link{vecchia_parent_structure}}.}

\item{threads}{Number of threads to use for parallelisation, taken from
\code{getOption('geowarp.threads')} by default, which is itself set to 1 by
default. The special value -1 picks a number of threads based on the number
of cores in the system).}
}
\value{
A vector containing the sampled alpha beta parameters.
}
\description{
This function gives the posterior distribution of the alpha beta parameters
in a fitted GeoWarp model.
}
