% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{stan_trust_optim}
\alias{stan_trust_optim}
\title{Find Optimal Parameters for a Stan Model Using Trust Region Optimisation}
\usage{
stan_trust_optim(
  object,
  data,
  report_level = 0,
  control = list(maxit = 2000, prec = 0.001),
  method = "SR1",
  initial_value_attempts = 10,
  initial_range = 4,
  theta_initial,
  ...
)
}
\arguments{
\item{object}{A Stan model object.}

\item{data}{A list containing the data to be used in the Stan model.}

\item{report_level}{The level of output to be produced by the optimisation.
Zero means no output, four gives the maximum.}

\item{control}{A list of control parameters passed to
\code{\link[trustOptim]{trust.optim}}.}

\item{method}{A character string indicating the method to be used in
\code{\link[trustOptim]{trust.optim}}.}

\item{initial_value_attempts}{The number of attempts to find a valid initial
value for the optimisation. Default is 10.}

\item{initial_range}{The range of the initial values for the optimisation on
the unconstrained space.}

\item{...}{Additional arguments passed to
\code{\link[trustOptim]{trust.optim}}.}
}
\value{
A list containing the following:
\itemize{
\item \code{value}: The optimized value of the objective function.
\item \code{par}: The optimized parameters in the original parameter space.
\item \code{iterations}: The number of iterations taken by the algorithm.
\item \code{status}: The status of the optimisation.
This mirrors the return structure of \code{\link[rstan]{optimizing}}.
}
}
\description{
This function optimizes the parameters of a given Stan model using trust
region optimisation. The function internally uses the
\code{\link[trustOptim]{trust.optim}}  function from the
\pkg{trustOptim} package and \code{\link[rstan]{log_prob}} and
\code{\link[rstan]{grad_log_prob}} from the \pkg{rstan} package.
}
\examples{
\dontrun{
  # Define a Stan model (assuming `model_code` contains your Stan model)
  stan_model <- rstan::stan_model(model_code = model_code)
  # Define data (assuming `data_list` contains your data)
  fit <- stan_trust_optim(stan_model, data = data_list)
}

}
\seealso{
\code{\link[trustOptim]{trust.optim}}
}
