% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia.R
\name{vecchia_grouping}
\alias{vecchia_grouping}
\title{Calculate Groups of Locations}
\usage{
vecchia_grouping(parent_structure, exponent = 3)
}
\arguments{
\item{parent_structure}{A parent structure list as returned by
\code{\link{vecchia_parent_structure}}.}

\item{exponent}{Criterion for grouping, passed to \code{GpGp::group_obs}. Should
be either 2 or 3.}
}
\value{
A list with the following entries:
\itemize{
\item \code{N_indices}: The length of \code{all_inds}.
\item \code{N_blocks}: The number of groups/blocks.
\item \code{block_indices}: A vector of which indices belong to which group.
Contains the indices in block 1 first, followed by those in block 2, and so
on. The indices in each block are ordered with the responses at the end.
\item \code{block_last_index}: For each block, the last index in
\code{block_indices} that contains the block's data indices.
\item \code{block_N_responses}: For each block, the number of responses in
the block.
}
}
\description{
Takes a parent structure and automatically partition the array into groups
that share many parents. In each group, some observations are conditioned on,
and some are the responses. No response ever belongs to more than one group.
This is used in GeoWarp's model fitting to calculate the model likelihood.
Under this hood this uses \code{\link[GpGp]{group_obs}} function in the GpGp
package.
}
\seealso{
\code{\link[GpGp]{group_obs}}
}
