% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia.R
\name{vecchia_parent_structure}
\alias{vecchia_parent_structure}
\title{Calculate a Parent Structure for the Vecchia Approximation}
\usage{
vecchia_parent_structure(
  observed_df,
  model,
  n_parents = 20,
  observed_ordering,
  observed_type = c("vertically_dense", "regular"),
  observed_options = list(n_parents = n_parents),
  prediction_df,
  prediction_type = c("regular", "vertically_dense"),
  prediction_ordering,
  prediction_within_options = list(n_parents = floor(n_parents/2)),
  prediction_between_options = list(n_parents = ceiling(n_parents/2))
)
}
\arguments{
\item{observed_df}{Data frame containing observed locations.}

\item{model}{GeoWarp model objects.}

\item{n_parents}{Number of parents to use. Default is 20.}

\item{observed_ordering}{Ordering of the observed locations. Defaults to
a random reordering, except for vertical only models.}

\item{observed_type}{Pattern of the observed locations. One of
'vertically_dense' (the default, useful for CPT data), or 'regular' (useful
for data on a grid or that is uniformly distributed).}

\item{observed_options}{List of additional options for observed structure.
See details below.}

\item{prediction_df}{Optional data frame containing prediction locations.}

\item{prediction_type}{Pattern of the predicted locations; see
\code{observed_type}. Defaults to regular, for a grid.}

\item{prediction_ordering}{Ordering of the prediction locations; see
\code{observed_ordering}.}

\item{prediction_within_options}{Options for the parent structure within
the prediction locations.}

\item{prediction_between_options}{Options for the parent structure between
the prediction and the observation locations.}
}
\value{
A list with the following components:
\itemize{
\item \code{observed_ordering}: A vector of indices specifying the ordering
used for observed locations.
\item \code{observed_parents}: The parent structure for observed locations.
Given as a matrix. The first column contains the index of the (reordered)
location, and the subsequent columns contain the indices of the parents (some
of which can be \code{NA}).
\item \code{prediction_ordering}: The ordering used for prediction locations
(if provided).
\item \code{prediction_within_parents}: Matrix of parents for within the
prediction locations (if provided).
\item \code{prediction_between_parents}: Matrix of parents for between the
predictions and observations.
}
}
\description{
This function calculates the parent structure for both observed and optional
predicted locations based on the Vecchia approximation. It's used to
structure the covariance or conditional dependencies in GeoWarp for both
inference and prediction. All observed locations are placed earlier in
the ordering than the predicted locations.
}
\details{
The chosen parent structure depends on the pattern of the observed and
prediction locations, which is user specified. The observed locations are
conditioned on each other, while the predicted locations condition on both
the observed and the other predicted locations to a degree specified by the
user.
}
\section{Location pattern and chosen parent structure}{

The observed data condition on themselves. For
\code{observed_type = 'vertically_dense'}, the scheme is to split the
parents, by default 50-50, between the nearest neighbours of the points among
its predecessors in the ordering, and points at other horizontal locations.
This split can be changed by setting \code{observed_options$n_parents} and
For \code{observed_type = 'regular'}, the parents are just the nearest
neighbours among the predecessor points. Then \code{observed_options} must contain
the number of parents in \code{n_parents} and, optionally, a vector named
\code{scaling} by which each coordinate will be rescaled; this can be used if the
grid is denser in one direction than another to ensure the neighbourhood
structure is regular too.

The prediction locations condition on both the observed locations and the
other prediction locations. The options for the scheme for conditioning
within the predictions is the same as for the observed locations, and the
behaviour of this scheme is controlled through \code{prediction_within_options},
which has the same options as \code{observed_options}.

For the between-parents, the scheme depends on \code{observed_type}: if the
observations are vertically dense, the observation parents are spread among
the different observed horizontal locations; otherwise, the nearest
neighbours among the parents are used. In both cases, the number of between
parents is controlled by \code{prediction_between_options$n_parents}. For
\code{observed_type = 'regular'}, the \code{scaling} option can be used for the
same effect as for the observed locations.
}

