% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warping.R
\name{warped_coordinates}
\alias{warped_coordinates}
\alias{unwarped_coordinates}
\title{Warped and Unwarp Coordinates}
\usage{
warped_coordinates(
  fit,
  df,
  model = fit$model,
  parameters = fit$parameters,
  x = model_coordinates(df, model)
)

unwarped_coordinates(
  fit,
  warped_x,
  model = fit$model,
  parameters = fit$parameters
)
}
\arguments{
\item{fit}{A fitted GeoWarp object.}

\item{df}{A dataframe containing the original coordinate space data.}

\item{model}{A GeoWarp model object, defaulting to \code{fit$model}.}

\item{parameters}{Parameters for the GeoWarp model, defaulting to
\code{fit$parameters}.}

\item{x}{The unwarped coordinates as a matrix, by default equal to
\code{model_coordinates(df, fit$model)}.}

\item{warped_x}{Matrix containing the warped coordinates to be transformed
back.}
}
\value{
A matrix where each row represents a point in the warped coordinate
space.
}
\description{
These function produces the warped coordinates for a GeoWarp model given the
original coordinates, or give the original coordinates given the warped
coordinates.
}
\section{Functions}{
\itemize{
\item \code{unwarped_coordinates()}: Calculated unwarped coordinates.

}}
\seealso{
\code{\link[geowarp]{geowarp_model}}
\code{\link[geowarp]{geowarp_optimize}}
\code{\link[geowarp]{model_coordinates}}
}
