% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geowarp-package.R
\docType{package}
\name{geowarp-package}
\alias{geowarp-package}
\alias{geowarp}
\title{The 'geowarp' package.}
\description{
A package to fit GeoWarp, a model that uses warped spatial
processes for inferring soil properties.
}
