% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{geowarp_optimise}
\alias{geowarp_optimise}
\title{Estimate the GeoWarp Model Parameters Through Optimisation}
\usage{
geowarp_optimise(
  df,
  model = geowarp_model(),
  max_attempts = 20,
  best_of = 5,
  algorithm = "BFGS",
  vecchia = "auto",
  n_parents = 20,
  parent_structure = vecchia_parent_structure(df, model, n_parents, ),
  grouping = vecchia_grouping(parent_structure),
  ...
)
}
\arguments{
\item{df}{Data frame containing the observations.}

\item{model}{A GeoWarp model object specifying the model structure and
priors,created using \code{\link{geowarp_model}}.}

\item{max_attempts}{Maximum number of optimisation attempts. Defaults to 20.}

\item{best_of}{Number of times to run the optimisation algorithm. The
algorithm is run repeatedly with different starting values in order to help
find the global posterior mode.}

\item{algorithm}{Optimisation algorithm to use, passed to
\code{\link[rstan]{optimizing}}.}

\item{vecchia}{Whether to use the Vecchia approximation. Can be 'auto',
\code{TRUE}, or \code{FALSE}. If 'auto', the Vecchia approximation is used if there
are more than 1000 observations.}

\item{n_parents}{Number of parent locations for the Vecchia approximation.
Defaults to 20.}

\item{parent_structure}{The parent structure used for the Vecchia
approximation; see \code{\link{vecchia_parent_structure}}.}

\item{grouping}{Grouping structure for the Vecchia approximation; see
\code{\link{vecchia_grouping}}. The default should be fine for most cases.}

\item{...}{Additional parameters to be passed to
\code{\link[rstan]{optimizing}}. You can use this to control the degree of
output.}
}
\value{
A 'geowarp_fit' object that estimated parameters, suitable for use
with \code{\link{predict.geowarp_fit}} and other methods.
}
\description{
This function estimates the parameters of a specified GeoWarp model by using
an optimisation algorithm to find the posterior mode.
}
\examples{
# Using the geowarp_optimise function
fit <- geowarp_optimise(
  df = my_data,
  model = my_model
)

}
